move loss to models
